Updated Border Style control in all blocks
