+++ src/blocks/count-down/components/inspector.js
@@ -15,6 +15,7 @@ import ResponsiveNewPaddingControl from "../../../settings-components/Responsive
 import ResponsiveNewMarginControl from "../../../settings-components/ResponsiveNewSpacingSettings/ResponsiveNewMarginControl/index";
 import RbeaRangeControl from "../../../utils/components/rbea-range-control";
 import RbeaColorControl from "../../../utils/components/rbea-color-control";
+import RbeaBorderStyleTabControl from "../../../utils/components/rbea-border-style-tab-control";
 
 // Setup the block
 const { __ } = wp.i18n;
@@ -490,48 +491,9 @@ export default class Inspector extends Component {
                 title={__("Border", "responsive-block-editor-addons")}
                 initialOpen={false}
               >
-                <SelectControl
-                  label={__("Border Style", "responsive-block-editor-addons")}
-                  value={boxBorderStyle}
+                <RbeaBorderStyleTabControl
+                  selected={boxBorderStyle}
                   onChange={(value) => setAttributes({ boxBorderStyle: value })}
-                  options={[
-                    {
-                      value: "none",
-                      label: __("None", "responsive-block-editor-addons"),
-                    },
-                    {
-                      value: "solid",
-                      label: __("Solid", "responsive-block-editor-addons"),
-                    },
-                    {
-                      value: "dotted",
-                      label: __("Dotted", "responsive-block-editor-addons"),
-                    },
-                    {
-                      value: "dashed",
-                      label: __("Dashed", "responsive-block-editor-addons"),
-                    },
-                    {
-                      value: "double",
-                      label: __("Double", "responsive-block-editor-addons"),
-                    },
-                    {
-                      value: "groove",
-                      label: __("Groove", "responsive-block-editor-addons"),
-                    },
-                    {
-                      value: "inset",
-                      label: __("Inset", "responsive-block-editor-addons"),
-                    },
-                    {
-                      value: "outset",
-                      label: __("Outset", "responsive-block-editor-addons"),
-                    },
-                    {
-                      value: "ridge",
-                      label: __("Ridge", "responsive-block-editor-addons"),
-                    },
-                  ]}
                 />
                 {"none" != boxBorderStyle && (
                   <Fragment>
