+++ src/blocks/how-to/components/inspector.js
@@ -17,6 +17,7 @@ import RbeaRangeControl from "../../../utils/components/rbea-range-control";
 import RbeaColorControl from "../../../utils/components/rbea-color-control";
 import RbeaMediaUploadControl from "../../../utils/components/rbea-media-upload-control";
 import RbeaTabRadioControl from "../../../utils/components/rbea-tab-radio-control";
+import RbeaBorderStyleTabControl from "../../../utils/components/rbea-border-style-tab-control";
 
 // Import block components
 const {
@@ -421,48 +422,9 @@ export default class Inspector extends Component {
                             />
                             {
                                 mainImage.url && (
-                                    <SelectControl
-                                        label={__("Border Style", "responsive-block-editor-addons")}
-                                        value={borderStyle}
+                                    <RbeaBorderStyleTabControl
+                                        selected={borderStyle}
                                         onChange={(value) => setAttributes({ borderStyle: value })}
-                                        options={[
-                                            {
-                                                value: "none",
-                                                label: __("None", "responsive-block-editor-addons"),
-                                            },
-                                            {
-                                                value: "solid",
-                                                label: __("Solid", "responsive-block-editor-addons"),
-                                            },
-                                            {
-                                                value: "dotted",
-                                                label: __("Dotted", "responsive-block-editor-addons"),
-                                            },
-                                            {
-                                                value: "dashed",
-                                                label: __("Dashed", "responsive-block-editor-addons"),
-                                            },
-                                            {
-                                                value: "double",
-                                                label: __("Double", "responsive-block-editor-addons"),
-                                            },
-                                            {
-                                                value: "groove",
-                                                label: __("Groove", "responsive-block-editor-addons"),
-                                            },
-                                            {
-                                                value: "inset",
-                                                label: __("Inset", "responsive-block-editor-addons"),
-                                            },
-                                            {
-                                                value: "outset",
-                                                label: __("Outset", "responsive-block-editor-addons"),
-                                            },
-                                            {
-                                                value: "ridge",
-                                                label: __("Ridge", "responsive-block-editor-addons"),
-                                            },
-                                        ]}
                                     />
                                 )
                             }
