+++ src/blocks/video-popup/components/inspector.js
@@ -11,6 +11,7 @@ import RbeaRangeControl from "../../../utils/components/rbea-range-control/index
 import RbeaColorControl from "../../../utils/components/rbea-color-control/index.js";
 import RbeaTabRadioControl from "../../../utils/components/rbea-tab-radio-control/index.js";
 import RbeaMediaUploadControl from "../../../utils/components/rbea-media-upload-control/index.js";
+import RbeaBorderStyleTabControl from "../../../utils/components/rbea-border-style-tab-control/index.js";
 
 
 // Setup the block
@@ -515,23 +516,11 @@ export default class Inspector extends Component {
                 title={__("Border", "responsive-block-editor-addons")}
                 initialOpen={false}
               >
-                <SelectControl
-                  label={__("Border Style", "responsive-block-editor-addons")}
-                  value={blockBorderStyle}
+                <RbeaBorderStyleTabControl
+                  selected={blockBorderStyle}
                   onChange={(value) =>
                     setAttributes({ blockBorderStyle: value })
                   }
-                  options={[
-                    { value: "none", label: __("None", "responsive-block-editor-addons") },
-                    { value: "solid", label: __("Solid", "responsive-block-editor-addons") },
-                    { value: "dotted", label: __("Dotted", "responsive-block-editor-addons") },
-                    { value: "dashed", label: __("Dashed", "responsive-block-editor-addons") },
-                    { value: "double", label: __("Double", "responsive-block-editor-addons") },
-                    { value: "groove", label: __("Groove", "responsive-block-editor-addons") },
-                    { value: "inset", label: __("Inset", "responsive-block-editor-addons") },
-                    { value: "outset", label: __("Outset", "responsive-block-editor-addons") },
-                    { value: "ridge", label: __("Ridge", "responsive-block-editor-addons") },
-                  ]}
                 />
                 {"none" != blockBorderStyle && (
                   <RbeaRangeControl
