--- src/settings-components/SectionBlockBorderSettings/editor.scss
@@ -1,19 +1,0 @@
-.rbea-border-style-selector {
-    label {
-        width: 100%;
-        height: 100%;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
-
-    .components-radio-control__option input:checked+label {
-        border: 1px solid #007CBA;
-        border-radius: 4px;
-        background-color: #F1FAFF;
-
-        .rbea-border-style-select-control-label{
-            color: #007CBA;
-        }
-    }
-}
