+++ src/utils/components/rbea-border-style-tab-control/styles/editor.scss
@@ -1,0 +1,52 @@
+.rbea-inspector-tab {
+    .rbea-border-style-selector {
+        label {
+            width: 100%;
+            height: 100%;
+            display: flex;
+            align-items: center;
+            justify-content: center;
+        }
+
+        .components-radio-control__option input:checked+label {
+            border: 1px solid #007CBA;
+            border-radius: 4px;
+            background-color: #F1FAFF;
+
+            .rbea-border-style-select-control-label{
+                color: #007CBA;
+            }
+        }
+    }
+
+    .rbea-border-style-selector {
+
+        .components-radio-control__option{
+        height: 35px;
+        width: 100%;
+        border: 1px solid #D9D9D9;
+        border: 1px solid #D9D9D9;
+        border-radius: 4px;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        }
+    
+        .components-flex.components-h-stack.components-v-stack{
+        display: grid;
+        grid-template-columns: repeat(4, 1fr);
+        row-gap: 0px; 
+        column-gap: 3px; 
+        }
+    
+        .rbea-border-style-select-control-label{
+        font-family: Roboto;
+        font-size: 13px;
+        font-weight: 400;
+        line-height: 16.41px;
+        text-align: center;
+        color: #666666;
+        margin-bottom: 0px;
+        }
+    }
+}
